refactor: split off management of python server
